Adjust 404.ts page - 10% Fixes #22
